requirements.txt
typer
rich
google-generativeai
GitPython
python-dotenv
pydantic
fastmcp
